+++ riscv/insts/dsl/disasms/base/LoadDisassembler.kt
@@ -11,7 +11,7 @@ object LoadDisassembler : InstructionDisassembler {
         val name = Instruction[mcode].name
         val rs1 = mcode[InstructionField.RS1]
         val rd = mcode[InstructionField.RD]
-        val imm = signExtend(mcode[InstructionField.IMM_11_0], 12)
+        val imm = signExtend(mcode[InstructionField.IMM_11_0].toInt(), 12)
         return "$name x$rd $imm(x$rs1)"
     }
 }
