+++ riscv/insts/dsl/disasms/extensions/FITypeDisassembler.kt
@@ -14,7 +14,7 @@ object FITypeDisassembler : InstructionDisassembler {
         val name = Instruction[mcode].name
         val rd = mcode[InstructionField.RD]
         val rs1 = mcode[InstructionField.RS1]
-        val imm = signExtend(mcode[InstructionField.IMM_11_0], 12)
+        val imm = signExtend(mcode[InstructionField.IMM_11_0].toInt(), 12)
         return "$name f$rd $imm(x$rs1)"
     }
 }
