+++ riscv/insts/dsl/formats/InstructionFormat.kt
@@ -7,7 +7,7 @@ data class FieldEqual(val ifield: InstructionField, val required: Int)
 
 open class InstructionFormat(val length: Int, val ifields: List<FieldEqual>) {
     fun matches(mcode: MachineCode): Boolean = ifields.all {
-        (ifield, required) -> mcode[ifield] == required
+        (ifield, required) -> mcode[ifield].toInt() == required
     }
 
     fun fill(): MachineCode {
