+++ riscv/insts/dsl/impls/base/STypeImplementation32.kt
@@ -9,19 +9,19 @@ import venusbackend.simulator.Simulator
 
 class STypeImplementation32(private val store: (Simulator, Int, Int) -> Unit) : InstructionImplementation {
     override operator fun invoke(mcode: MachineCode, sim: Simulator) {
-        val rs1 = mcode[InstructionField.RS1]
-        val rs2 = mcode[InstructionField.RS2]
+        val rs1 = mcode[InstructionField.RS1].toInt()
+        val rs2 = mcode[InstructionField.RS2].toInt()
         val imm = constructStoreImmediate(mcode)
-        val vrs1 = sim.getReg(rs1)
-        val vrs2 = sim.getReg(rs2)
+        val vrs1 = sim.getReg(rs1).toInt()
+        val vrs2 = sim.getReg(rs2).toInt()
         store(sim, vrs1 + imm, vrs2)
         sim.incrementPC(mcode.length)
     }
 }
 
 fun constructStoreImmediate(mcode: MachineCode): Int {
-    val imm_11_5 = mcode[InstructionField.IMM_11_5]
-    val imm_4_0 = mcode[InstructionField.IMM_4_0]
+    val imm_11_5 = mcode[InstructionField.IMM_11_5].toInt()
+    val imm_4_0 = mcode[InstructionField.IMM_4_0].toInt()
     var imm = 0
     imm = setBitslice(imm, imm_11_5, 5, 12)
     imm = setBitslice(imm, imm_4_0, 0, 5)
