+++ riscv/insts/dsl/impls/base/ShiftImmediateImplementation32.kt
@@ -7,10 +7,10 @@ import venusbackend.simulator.Simulator
 
 class ShiftImmediateImplementation32(private val eval: (Int, Int) -> Int) : InstructionImplementation {
     override operator fun invoke(mcode: MachineCode, sim: Simulator) {
-        val rs1 = mcode[InstructionField.RS1]
-        val shamt = mcode[InstructionField.SHAMT]
-        val rd = mcode[InstructionField.RD]
-        val vrs1 = sim.getReg(rs1)
+        val rs1 = mcode[InstructionField.RS1].toInt()
+        val shamt = mcode[InstructionField.SHAMT].toInt()
+        val rd = mcode[InstructionField.RD].toInt()
+        val vrs1 = sim.getReg(rs1).toInt()
         sim.setReg(rd, eval(vrs1, shamt))
         sim.incrementPC(mcode.length)
     }
