+++ riscv/insts/dsl/impls/extensions/FITypeImplementation32.kt
@@ -12,10 +12,10 @@ import venusbackend.simulator.Simulator
  */
 class FITypeImplementation32(private val eval: (Int, Simulator) -> Decimal) : InstructionImplementation {
     override operator fun invoke(mcode: MachineCode, sim: Simulator) {
-        val rs1 = mcode[InstructionField.RS1]
-        val rd = mcode[InstructionField.RD]
-        val vrs1 = sim.getReg(rs1)
-        val imm = signExtend(mcode[InstructionField.IMM_11_0], 12)
+        val rs1 = mcode[InstructionField.RS1].toInt()
+        val rd = mcode[InstructionField.RD].toInt()
+        val vrs1 = sim.getReg(rs1).toInt()
+        val imm = signExtend(mcode[InstructionField.IMM_11_0].toInt(), 12)
         sim.setFReg(rd, eval(vrs1 + imm, sim))
         sim.incrementPC(mcode.length)
     }
