+++ riscv/insts/dsl/impls/extensions/FR4TypeImplementation32.kt
@@ -11,10 +11,10 @@ import venusbackend.simulator.Simulator
  */
 class FR4TypeImplementation32(private val eval: (Decimal, Decimal, Decimal) -> Decimal) : InstructionImplementation {
     override operator fun invoke(mcode: MachineCode, sim: Simulator) {
-        val rs1 = mcode[InstructionField.RS1]
-        val rs2 = mcode[InstructionField.RS2]
-        val rs3 = mcode[InstructionField.RS3]
-        val rd = mcode[InstructionField.RD]
+        val rs1 = mcode[InstructionField.RS1].toInt()
+        val rs2 = mcode[InstructionField.RS2].toInt()
+        val rs3 = mcode[InstructionField.RS3].toInt()
+        val rd = mcode[InstructionField.RD].toInt()
         val vrs1 = sim.getFReg(rs1)
         val vrs2 = sim.getFReg(rs2)
         val vrs3 = sim.getFReg(rs3)
