+++ riscv/insts/dsl/impls/extensions/RtFTypeImplementation32.kt
@@ -8,9 +8,9 @@ import venusbackend.simulator.Simulator
 
 class RtFTypeImplementation32(private val eval: (Int) -> Decimal) : InstructionImplementation {
     override operator fun invoke(mcode: MachineCode, sim: Simulator) {
-        val rs1 = mcode[InstructionField.RS1]
-        val rd = mcode[InstructionField.RD]
-        val vrs1 = sim.getReg(rs1)
+        val rs1 = mcode[InstructionField.RS1].toInt()
+        val rd = mcode[InstructionField.RD].toInt()
+        val vrs1 = sim.getReg(rs1).toInt()
         sim.setFReg(rd, eval(vrs1))
         sim.incrementPC(mcode.length)
     }
